Split raw MIME offload test scenarios
